app/requirements.txt
aiofile==3.8.7
aiohttp==3.8.4
aiosignal==1.3.1
aiotdlib==0.19.2
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
caio==0.9.12
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2
exceptiongroup==1.1.2
executing==1.2.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==5.3.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
wcwidth==0.2.6
yarl==1.9.2
mixpanel==4.10.0
celery==5.2.7
flower==1.0.0
redis==4.3.3
django-celery-beat==2.5.0
